+++ vdifunctions.h
@@ -24,7 +24,7 @@ void addBitsFromStreamData(std::vector<bool> *bits, int numBits, long long seek_
 unsigned long long convertEndian(unsigned char C[], int size, bool littleEndian = false);
 unsigned char getCharFromStream(int size, long long seek_to, std::ifstream &input);
 unsigned long long getStreamData(int size, long long seek_to, std::ifstream &input, std::string name = "", bool output = true, bool littleEndian = false, bool afterVDIMap = true);
-QString FileSizeToString(unsigned long long size);
+QString FileSizeToString(unsigned int size);
